🦜🕹️  had a bath after much dancing
Surely types are a good idea, right? I mean. NOoooooooo!
The ugly ugly types. But I guess they are good.

Removing the lint from the belly button of a parrot is an impossible
task, because parrots don't have belly buttons. AND YET! This is
something that happened. Parrotjoy is lint free.